# backend/swagger.yaml
openapi: 3.0.0
info:
  title: dayNoteApp API (Obsidian & Claude AI Edition)
  description: API specification for dayNoteApp including Authentication, Notes, Categories, Obsidian Knowledge Graph, MindMap, and Claude AI Assistant.
  version: 2.0.0
servers:
  - url: http://localhost:5000
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /api/notes:
    get:
      summary: Get user notes list
      parameters:
        - in: query
          name: category
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of notes
  /api/notes/{note_id}:
    get:
      summary: Get detailed note content by ID
      parameters:
        - in: path
          name: note_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detailed note with tags and wikilinks
    delete:
      summary: Delete a note
      parameters:
        - in: path
          name: note_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Note deleted successfully
  /api/notes/{note_id}/backlinks:
    get:
      summary: Get backlinks referencing this note
      parameters:
        - in: path
          name: note_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Linked and unlinked backlink references
  /api/notes/{note_id}/save:
    post:
      summary: Save note text content
      parameters:
        - in: path
          name: note_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
      responses:
        '200':
          description: Note content updated successfully
  /api/upload:
    post:
      summary: Upload a physical file note
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                category:
                  type: string
      responses:
        '201':
          description: File uploaded successfully
  /api/ai/upload:
    post:
      summary: Create an AI-generated note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
                extension:
                  type: string
      responses:
        '201':
          description: AI Note created
  /api/notes/url:
    post:
      summary: Save a web URL note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                url:
                  type: string
                category:
                  type: string
      responses:
        '201':
...